--- .github/workflows/validate-schema.yml
@@ -1,47 +1,0 @@
-name: "Validate Schema"
-
-on:
-  workflow_dispatch:
-    inputs:
-      stateId:
-        description: "State Id"
-      eventName:
-        description: "Event Name"
-      eventPayload:
-        description: "Event Payload"
-      settings:
-        description: "Settings"
-      authToken:
-        description: "Auth Token"
-      ref:
-        description: "Ref"
-
-jobs:
-  validate:
-    name: "Validate Schema"
-    runs-on: ubuntu-latest
-    permissions: write-all
-
-    steps:
-      - uses: actions/checkout@v4
-
-      - name: Setup node
-        uses: actions/setup-node@v4
-        with:
-          node-version: "20.10.0"
-
-      - name: Install deps and run validation
-        run: |
-          yarn install --immutable --immutable-cache --check-cache
-          yarn tsx src/validate-schema.ts
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          OPENAI_API_KEY: ${{ secrets.OPENAI_API_KEY }}
-          X25519_PRIVATE_KEY: ${{ secrets.X25519_PRIVATE_KEY }}
-          SUPABASE_URL: ${{ secrets.SUPABASE_URL }}
-          SUPABASE_KEY: ${{ secrets.SUPABASE_KEY }}
-          NFT_MINTER_PRIVATE_KEY: ${{ secrets.NFT_MINTER_PRIVATE_KEY }}
-          NFT_CONTRACT_ADDRESS: ${{ secrets.NFT_CONTRACT_ADDRESS }}
-          PERMIT_FEE_RATE: ${{ secrets.PERMIT_FEE_RATE }}
-          PERMIT_TREASURY_GITHUB_USERNAME: ${{ secrets.PERMIT_TREASURY_GITHUB_USERNAME }}
-          PERMIT_ERC20_TOKENS_NO_FEE_WHITELIST: ${{ secrets.PERMIT_ERC20_TOKENS_NO_FEE_WHITELIST }}
