+++ package.json
@@ -78,7 +78,7 @@
     "npm-run-all": "4.1.5",
     "prettier": "3.2.5",
     "ts-node": "10.9.2",
-    "typescript": "5.3.3"
+    "typescript": "5.6.2"
   },
   "lint-staged": {
     "*.ts": [
