+++ src/helpers/validator.ts
@@ -5,6 +5,8 @@ import {
   validateIncentivesConfiguration,
 } from "../configuration/incentives";
 import envConfigSchema, { EnvConfigType, envValidator } from "../types/env-type";
+import * as github from "@actions/github";
+import { Octokit } from "@octokit/rest";
 
 export function validateAndDecodeSchemas(rawEnv: object, rawSettings: object) {
   const errors: ValueError[] = [];
@@ -40,4 +42,22 @@ export function validateAndDecodeSchemas(rawEnv: object, rawSettings: object) {
     }
     throw e;
   }
+}
+
+export async function returnDataToKernel(
+  repoToken: string,
+  stateId: string,
+  output: object,
+  eventType = "return_data_to_ubiquibot_kernel"
+) {
+  const octokit = new Octokit({ auth: repoToken });
+  return octokit.repos.createDispatchEvent({
+    owner: github.context.repo.owner,
+    repo: github.context.repo.repo,
+    event_type: eventType,
+    client_payload: {
+      state_id: stateId,
+      output: JSON.stringify(output),
+    },
+  });
 }
