+++ src/run.ts
@@ -2,13 +2,14 @@ import configuration from "./configuration/config-reader";
 import githubCommentModuleInstance from "./helpers/github-comment-module-instance";
 import { getSortedPrices } from "./helpers/label-price-extractor";
 import logger from "./helpers/logger";
+import { returnDataToKernel } from "./helpers/validator";
 import { IssueActivity } from "./issue-activity";
 import program from "./parser/command-line";
 import { Processor } from "./parser/processor";
 import { parseGitHubUrl } from "./start";
 
 export async function run() {
-  const { eventPayload, eventName } = program;
+  const { eventPayload, eventName, stateId } = program;
   if (eventName === "issues.closed") {
     if (eventPayload.issue.state_reason !== "completed") {
       return logger.info("Issue was not closed as completed. Skipping.").logMessage.raw;
@@ -23,7 +24,9 @@ export async function run() {
     }
     const processor = new Processor();
     await processor.run(activity);
-    return processor.dump();
+    const result = processor.dump();
+    await returnDataToKernel(process.env.GITHUB_TOKEN, stateId, { result });
+    return result;
   } else {
     return logger.error(`${eventName} is not supported, skipping.`).logMessage.raw;
   }
